Refactor ESP32 NimBLE backend
